Refactor: SocialMediaBadge to SocialPostBadge
