--- addon/components/o-s-s/social-media-badge.hbs
@@ -1,8 +1,0 @@
-<div class={{this.computedClasses}} role="button"
-     {{enable-tooltip title=@tooltip}}
-     {{on "click" this.onMediaTypeClick}}
-     ...attributes>
-    <div class="icon-container">
-        <i class={{this.iconDefinition}}></i>
-    </div>
-</div>
