+++ addon/components/o-s-s/social-post-badge.hbs
@@ -1,0 +1,7 @@
+<div class={{this.computedClasses}} role="button"
+     {{on "click" this.onPostTypeClick}}
+     ...attributes>
+  <div class="icon-container">
+    <i class={{this.iconDefinition}}></i>
+  </div>
+</div>
