--- app/components/o-s-s/social-media-badge.js
@@ -1,1 +1,0 @@
-export { default } from '@upfluence/oss-components/components/o-s-s/social-media-badge';
