+++ app/components/o-s-s/social-post-badge.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/oss-components/components/o-s-s/social-post-badge';
