+++ app/styles/oss-components.less
@@ -30,7 +30,7 @@
 @import 'atoms/input-container';
 @import 'atoms/input-group';
 @import 'atoms/tooltip';
-@import 'atoms/social-media-badge';
+@import 'atoms/social-post-badge';
 @import 'molecules/select';
 @import 'molecules/nav-tab';
 @import 'molecules/button-dropdown';
